Update Palindrome Linked List
In Palindromes Land, every linked list is a palindrome. You are a code policeman in Palindromes Land, and you need to make sure that only those linked lists that are Palindromes can enter the city.

You are given a linked list that has n nodes representing a linked list in Palindromes Land. Print 1 if it is a palindrome and 0, otherwise